# backend-go/api/openapi.yaml
openapi: 3.0.3
info:
  title: Creation Agent Studio API
  version: 2.0.0
  description: |
    Creation Agent Studio backend contract (Go control plane).

    This file is the single source of truth for the HTTP API.
    Go server types and TypeScript clients are generated from it.

    ## Error envelope conventions (legacy-compatible)

    The React frontend was validated against the Django reference
    implementation, which produced four distinct error shapes:

    - `{"detail": "message"}` — 403/404 style errors (DRF PermissionDenied / NotFound).
    - `{"error": "message"}` — simple JSON errors (identity endpoints, 409/502 conflicts).
    - `{"field": ["message", ...]}` — field validation errors (DRF serializer style).
    - `["message"]` — bare array, 400 for hand-raised DRF ValidationErrors (e.g. "application_id is required").

    All four shapes are reproduced exactly. Clients should read the
    `detail` / `error` / first field message in that order.

    ## Authentication

    Normal users authenticate with the HttpOnly opaque session cookie
    `studio_session` (set by OAuth exchange / admin login). All
    state-changing requests additionally require the CSRF header
    `X-CSRF-Token` matching the `studio_csrf` cookie (double-submit).
    `GET /api/v2/artifacts/{id}/open` is cookie-authenticated because it is
    opened via browser navigation (window.open) and cannot send headers.

    ## SSE

    `GET /api/v2/runs/{id}/stream` responds with `text/event-stream`.
...
servers:
...
paths:
...
  /api/v2/feishu/forward/targets:
    get:
      operationId: listFeishuForwardTargets
      tags: [runs]
      summary: Feishu forward targets (contacts or groups) for the caller
      parameters:
        - name: type
          in: query
          required: true
          schema: { type: string, enum: [user, chat] }
          description: user = org directory search, chat = the caller's groups.
        - name: query
          in: query
          schema: { type: string }
          description: Name search (user type only).
      responses:
        "200":
          description: Bare array of targets
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/FeishuForwardTarget" }
        "400":
          description: caller has no bound Feishu identity / bad type
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
        "401":
          description: not authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
  /api/v2/feishu/forward:
    post:
      operationId: forwardShareToFeishu
      tags: [runs]
      summary: Deliver a share snapshot card to Feishu users/groups (owner only)
      description: Sends an interactive card carrying the share link. The link
        origin comes from the request Origin header. Per-target results are
        returned; a failing user-token send falls back to the app token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [share_token, targets]
              properties:
                share_token: { type: string }
                targets:
                  type: array
                  minItems: 1
                  maxItems: 20
                  items:
                    type: object
                    required: [target_type, id]
                    properties:
                      target_type: { type: string, enum: [user, chat] }
                      id: { type: string }
      responses:
        "200":
          description: Per-target delivery results
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FeishuForwardResult" }
        "400":
          description: invalid body / share not owned by caller
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
        "401":
          description: not authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }

  # ─────────────────────────────────────────────────────────────── auth ──
...
components:
...
  schemas:
...
    FeishuForwardTarget:
      type: object
      required: [id, name, target_type]
      properties:
        id: { type: string }
        name: { type: string }
        avatar_url: { type: string }
        target_type: { type: string, enum: [user, chat] }
    FeishuForwardResult:
      type: object
      required: [results, success_count, fail_count]
      properties:
        results:
          type: array
          items:
            type: object
            required: [target_id, ok]
            properties:
              target_id: { type: string }
              ok: { type: boolean }
              error: { type: string }
        success_count: { type: integer }
        fail_count: { type: integer }
...
    DetailError:
      type: object
      required: [detail]
      properties:
        detail: { type: string }
...